refactor(storage): move `Rest*` to a subdirectory
