+++ google/cloud/storage/google_cloud_cpp_storage.bzl
@@ -99,9 +99,9 @@ google_cloud_cpp_storage_hdrs = [
     "internal/raw_client.h",
     "internal/raw_client_wrapper_utils.h",
     "internal/request_project_id.h",
-    "internal/rest_client.h",
-    "internal/rest_object_read_source.h",
-    "internal/rest_request_builder.h",
+    "internal/rest/client.h",
+    "internal/rest/object_read_source.h",
+    "internal/rest/request_builder.h",
     "internal/retry_client.h",
     "internal/retry_object_read_source.h",
     "internal/service_account_parser.h",
@@ -219,9 +219,9 @@ google_cloud_cpp_storage_srcs = [
     "internal/policy_document_request.cc",
     "internal/raw_client.cc",
     "internal/request_project_id.cc",
-    "internal/rest_client.cc",
-    "internal/rest_object_read_source.cc",
-    "internal/rest_request_builder.cc",
+    "internal/rest/client.cc",
+    "internal/rest/object_read_source.cc",
+    "internal/rest/request_builder.cc",
     "internal/retry_client.cc",
     "internal/retry_object_read_source.cc",
     "internal/service_account_parser.cc",
