+++ google/cloud/storage/google_cloud_cpp_storage.cmake
@@ -167,12 +167,12 @@ add_library(
     internal/raw_client_wrapper_utils.h
     internal/request_project_id.cc
     internal/request_project_id.h
-    internal/rest_client.cc
-    internal/rest_client.h
-    internal/rest_object_read_source.cc
-    internal/rest_object_read_source.h
-    internal/rest_request_builder.cc
-    internal/rest_request_builder.h
+    internal/rest/client.cc
+    internal/rest/client.h
+    internal/rest/object_read_source.cc
+    internal/rest/object_read_source.h
+    internal/rest/request_builder.cc
+    internal/rest/request_builder.h
     internal/retry_client.cc
     internal/retry_client.h
     internal/retry_object_read_source.cc
@@ -506,9 +506,9 @@ if (BUILD_TESTING)
         internal/patch_builder_test.cc
         internal/policy_document_request_test.cc
         internal/request_project_id_test.cc
-        internal/rest_client_test.cc
-        internal/rest_object_read_source_test.cc
-        internal/rest_request_builder_test.cc
+        internal/rest/client_test.cc
+        internal/rest/object_read_source_test.cc
+        internal/rest/request_builder_test.cc
         internal/retry_client_test.cc
         internal/retry_object_read_source_test.cc
         internal/service_account_requests_test.cc
