+++ google/cloud/storage/grpc_plugin_test.cc
@@ -15,7 +15,7 @@
 #include "google/cloud/storage/grpc_plugin.h"
 #include "google/cloud/storage/internal/grpc/client.h"
 #include "google/cloud/storage/internal/hybrid_client.h"
-#include "google/cloud/storage/internal/rest_client.h"
+#include "google/cloud/storage/internal/rest/client.h"
 #include "google/cloud/storage/internal/retry_client.h"
 #include "google/cloud/storage/options.h"
 #include "google/cloud/common_options.h"
