+++ google/cloud/storage/internal/rest/object_read_source.h
