+++ google/cloud/storage/internal/rest/request_builder.h
