+++ google/cloud/storage/storage_client_unit_tests.bzl
@@ -74,9 +74,9 @@ storage_client_unit_tests = [
     "internal/patch_builder_test.cc",
     "internal/policy_document_request_test.cc",
     "internal/request_project_id_test.cc",
-    "internal/rest_client_test.cc",
-    "internal/rest_object_read_source_test.cc",
-    "internal/rest_request_builder_test.cc",
+    "internal/rest/client_test.cc",
+    "internal/rest/object_read_source_test.cc",
+    "internal/rest/request_builder_test.cc",
     "internal/retry_client_test.cc",
     "internal/retry_object_read_source_test.cc",
     "internal/service_account_requests_test.cc",
